% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_v_celltype_distance.R
\name{condition_v_celltype_distance}
\alias{condition_v_celltype_distance}
\title{Condition vs Cell-type distances}
\usage{
condition_v_celltype_distance(
  normalized_counts,
  meta.data,
  fixed.effects,
  random.effects = c(),
  clusters,
  d = 20,
  truncate = FALSE,
  min.counts.per.cell = 20,
  weights = NULL
)
}
\arguments{
\item{normalized_counts}{A matrix containing
normalized data with genes on rows and cells
on columns}

\item{meta.data}{A data frame containing meta data for each cell.}

\item{fixed.effects}{The columns in meta.data corresponding to the fixed effects. In
a typical case, this would be the condition of interest.}

\item{random.effects}{The columns in meta.data corresponding to the random effects.
In a typical use case this would be the column containing patient ID.}

\item{clusters}{The column containing the cell-type annotation.}

\item{d}{The number of PCs to use.}

\item{truncate}{Whether or not to round negative distances to 0.}

\item{min.counts.per.cell}{The minimum number of cells per cluster to perform the estimation.}

\item{weights}{An optional vector of length equal to the number of genes specifying the weight
to place on each gene in the distance estimate.}
}
\value{
A list with components
\itemize{
\item \code{p} - A boxplot comparing the condition perturbation to the
inter-celltype distances.
\item \code{D} - A distance matrix for the cell types.
}
}
\description{
Compare the condition perturbation to the estimated distance between
different cell types
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
