% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{distGenes}
\alias{distGenes}
\title{Plot the direction of the perturbation}
\usage{
distGenes(scd.object, cluster)
}
\arguments{
\item{scd.object}{A list obtained from applying the main function \code{\link{scDist}}.}

\item{cluster}{The cluster to make the plot for}
}
\value{
A `ggplot2` object containing the plot
}
\description{
Plot the distance estimates and corresponding standard
errors
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
