% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBetas}
\alias{plotBetas}
\title{Plot the estimate difference in each PC.}
\usage{
plotBetas(scd.object, cluster)
}
\arguments{
\item{scd.object}{A list obtained from applying the main function \code{\link{scDist}}.}

\item{cluster}{The cluster to make the plot for}
}
\description{
Plots the estimated change in each PC.
}
\details{
scDist fits a linear model to each PC direction. The effect
of interest is plotted (as a function of PC) with this function.
}
