% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_analysis.R
\name{scDist_tree}
\alias{scDist_tree}
\title{Tree-based analysis of scDist distances}
\usage{
scDist_tree(sco)
}
\arguments{
\item{sco}{A `seurat` object containing the
following columns in meta data: `cellType`,
`condition`, and patient.}
}
\description{
Finds a cluster tree and then uses
the scDist procedure to estimate the distance at
each node of the tree
}
