% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_data.R
\name{simData}
\alias{simData}
\title{Simulate data according to the scDist model}
\usage{
simData(nct = 10, J = 50, N1, N2, G = 1000, nn = 100, tau = 0.5)
}
\arguments{
\item{nct}{The number of cell types.}

\item{J}{The number of cells per cell type.}

\item{N1}{The number of patients in first condition (reference).}

\item{N2}{The number of patients in second condition.}

\item{G}{The number of genes.}

\item{nn}{The number of non-null genes between the conditions.}

\item{tau}{The standard deviation of the patient-specific effect.}
}
\value{
A list with components
\itemize{
\item \code{Y} - A matrix containing normalized counts (genes x cells).
\item \code{meta.data} - Metadata that can be used as input to `scDist`.
\item \code{D.true} - The true distance
}
}
\description{
Simulate data according to the scDist model
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
