% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{DistPlot}
\alias{DistPlot}
\title{Plot the results of scDist}
\usage{
DistPlot(scd.object, return.plot = FALSE)
}
\arguments{
\item{scd.object}{A list obtained from applying the main function \code{\link{scDist}}.}
}
\value{
If return.plot is true, a ggplot object is returned. Otherwise,
nothing is returned.
}
\description{
Plot the distance estimates and corresponding standard
errors
}
\author{
Phillip B. Nicol <philnicol740@gmail.com>
}
