% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{FDRDistPlot}
\alias{FDRDistPlot}
\title{FDRDistPlot}
\usage{
FDRDistPlot(scd.object)
}
\arguments{
\item{scd.object}{A list obtained from applying the main function \code{\link{scDist}}.}
}
\description{
ADD DESCRIPTION HERE
}
