% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotBetas}
\alias{plotBetas}
\title{plotBetas}
\usage{
plotBetas(scd.object, cluster)
}
\arguments{
\item{scd.object}{A list obtained from applying the main function \code{\link{scDist}}.}

\item{cluster}{The cluster to make the plot for}
}
\description{
ADD DESCRIPTION HERE
}
